+++ pages/_app.js
@@ -17,7 +17,9 @@ function MyApp({ Component, pageProps }) {
       <Head>
         <title>{BANK_NAME}</title>
       </Head>
-      <Toaster position="top-right" richColors closeButton visibleToasts={9} />
+      <Toaster position="top-right" richColors closeButton visibleToasts={9} toastOptions={{
+        duration: 10000
+      }} />
       <PageLayout>
         <DockBankHelper />
